commit all layout & apis working & add questions last layout by islam emam

+++ SWE2_Project/src/main/java/swe2/springbootstarter/question/QuestionBaseRepository.java
@@ -10,7 +10,7 @@ import swe2.springbootstarter.entities.Question;
 
 
 @NoRepositoryBean
-public interface QuestionBaseRepository<T extends Question> extends CrudRepository<T, Long> {
+public interface QuestionBaseRepository<T extends Question> extends CrudRepository<T, Integer> {
 
 	public List<T> findByGameId(Integer gameName);
